fix(security): make Curve448 carry narrowing explicit
Replaces implicit narrowing compound assignments in Curve448 with explicit int casts to satisfy java/implicit-cast-in-compound-assignment without changing arithmetic behavior. Updates CHANGELOG Unreleased/Fixed.

+++ app/src/main/java/com/cirabit/android/noise/southernstorm/crypto/Curve448.java
@@ -220,11 +220,11 @@ public final class Curve448 {
 
 	    accum0 += accum8 + c[8];
 	    c[8] = ((int)accum0) & mask;
-	    c[9] += accum0 >>> 28;
+	    c[9] += (int)(accum0 >>> 28);
 
 	    accum8 += c[0];
 	    c[0] = ((int)accum8) & mask;
-	    c[1] += accum8 >>> 28;
+	    c[1] += (int)(accum8 >>> 28);
 	}
 
 	// p448_weak_reduce
